requirements.txt
Flask==2.3.3
APScheduler==3.10.4
google-generativeai==0.4.1
python-dotenv==1.0.1
supabase==2.3.5
httpx==0.27.0
playwright==1.42.0
